added Graphics/LayersMixingHiRes test (as Layer2Colours, but with HiRes)
This is similar to Layer2Colours test, but the ULA layer is
reconfigured into Timex "HiRes" 512x192 monochromatic mode.

The border is expected to be "pink" on real board (HiRes mode will
override it with paper colour, which is "10" (red 2 + 8 for bright) and
palette entry 128+10 is transparent pink, which should fallback into
fallback-pink (going by pure theory and HW pipeline explanation from Jim
Bagley on forums - needs test with real board to confirm).