added Sprites/Relative test, unfinished
The test is written according to current work-in-progress specifications
for core 2.00.27 or later, so there's at this moment no "Next" machine
capable to produce correct output (and there may be bugs in the code).

Although the code is now almost finished (maybe adding clip window would
be nice) - only verification and screenshot of correct result are
missing.