ULA/ClassicPaletized: enforcing +3 timing, ReadMe: timing warning added
The timing of this demo is not perfect, and will stay as that for the
moment, as the "+3/Next" timings itself seems to be not completely final
and correct (it's still quite likely the test timing is also wrong, and
will need extra round of fixing/tuning to get desired output).

The test is still valid even if not perfectly aligned, so added to the
suite any way.