Misc/ZilogDMA: added photos of real Zilog DMA results, refreshed readme/comments
The Next board photo shows the correct "DMA" letters in background, but
this was achieved by patched test doing second LOAD after correct
transfer direction is set. The final version of test does use only
single LOAD with the opposite direction (to load "fixed" port 254
address of Port B on Zilog DMA, otherwise it will be lost when
destination port is in "fixed" mode), so the top border with current
core 3.0.5 will show "noise" instead of the letters.

(the test works as desired with UA858D or Zilog DMA chip)