Misc/ZilogDMA: changing the graphics, colors tuning, TBBlue detection
Adjusting to new knowledge from testing with real HW, it turned out the
LOAD after transfer-direction change is quite essential even on Zilog
DMA, without the second LOAD the transfer (letters in top border) start
with single wrong byte extra, and stop one byte short before the end.

The new layout is modified to make the whole top-border transfer visible
on most of the TVs, so even the first wrong byte is visible.

Colour scheme was changed to not contain red color as long as the test
works correctly (the `otir` sequences are now marked with yellow border
blocks).

Emulators without the correct quirk behaviour to do the top border
transfer in "wrong" way (and do the "correct" transfer) will display the
first `otir` sequence for flashing block as "red".

The DMA port auto-detection was still unreliable, so rewritten instead
to detect TBBlue or Emulator and switch to $6B port as default there.