base/Copper: minor change in comment and order of NR $61/$62 setup
It's like third time I did think it's important to first stop Copper (by
writing to $62) before writing to $61, but it's bogus.

The $61 and $62 are write-to index registers, not "CPC" (Copper PC),
i.e. writing to $61 does not affect currently running Copper code, it
affects only the next write to Copper memory.

Still doing STOP first in the code, just to not derail me fourth time
when I will see the source after some time.