base/NextReg_defaults: fix write+verify of $06 "peripherals 2"
The keyboard/mouse bit can be written only during config-mode, so it
won't bulge during the test and either 0 or 1 has to be accepted
(depending on original value).

Also PI GPIO $98-$9B test is weakened to "any read" constraint, as there
are some readings on real Next with Pi0 attached.